+++ utils/intentmetrics.py
@@ -20,29 +20,44 @@ import csv
 import pandas as pd
 from argparse import ArgumentParser
 from sklearn.metrics import precision_recall_fscore_support
+from __init__ import UTF_8
 
 
 def func(args):
     in_df = pd.read_csv(args.in_file, quoting=csv.QUOTE_ALL,
-                        encoding='utf-8', keep_default_na=False)
+                        encoding=UTF_8, keep_default_na=False)
     # Look for target columns
     if args.test_column not in in_df or args.golden_column not in in_df:
         raise ValueError('Missing required columns')
 
     labels = in_df[args.golden_column].drop_duplicates().sort_values()
 
-    precision, recall, _, support = \
+    precisions, recalls, _, support = \
         precision_recall_fscore_support(y_true=in_df[args.golden_column],
                                         y_pred=in_df[args.test_column],
                                         labels=labels)
 
+    if args.partial_credit_on is not None:
+        for idx, label in enumerate(labels):
+            retrieved_doc_indx = in_df[args.test_column] == label
+            relevant_doc_indx = in_df[args.golden_column] == label
+            retrieved_doc_num = len(in_df[retrieved_doc_indx])
+            relevant_doc_num  = len(in_df[relevant_doc_indx])
+
+            precision = in_df[retrieved_doc_indx]['score'].sum() / retrieved_doc_num
+            recall = in_df[relevant_doc_indx]['score'].sum()/ relevant_doc_num
+
+            precisions[idx] = precision
+            recalls[idx] = recall
+
     out_df = pd.DataFrame(data={'intent': labels,
-                                'true positive rate': recall,
-                                'positive predictive value': precision,
+                                'true positive rate': recalls,
+                                'positive predictive value': precisions,
                                 'number of samples': support})
 
     out_df.to_csv(args.out_file, encoding='utf-8', quoting=csv.QUOTE_ALL,
-                  index=False)
+                  index=False, columns=['intent', 'number of samples', 'true positive rate',
+                  'positive predictive value'] )
 
 
 def create_parser():
@@ -59,6 +74,8 @@ def create_parser():
     parser.add_argument('-g', '--golden_column', type=str,
                         default='golden intent',
                         help='Golden column name')
+    parser.add_argument('-p', '--partial_credit_on', type=str,
+                        help='Use only if partial credit scoring ')
     return parser